Move data collection to a separate library
