+++ p2p.py
@@ -1,5 +1,4 @@
 import os
-import datetime
 import pandas
 from marketplace import mintos
 from marketplace import iuvo
@@ -7,6 +6,7 @@ from marketplace import marketplace
 from reports import diversification
 from reports import overall
 from logger import logger
+from datacollection.datacollection import collect_investment_data
 
 
 RELEVANT_COLUMNS = [marketplace.COUNTRY, marketplace.LOAN_ORIGINATOR, marketplace.OUTSTANDING_PRINCIPAL]
@@ -15,24 +15,6 @@ DATA_DIRECTORY = os.path.join('.', 'data')
 
 def get_latest_report_date(marketplace_files):
     return max(marketplace_files.keys())
-
-
-def read_marketplace_files(data_directory, investment_platform):
-    marketplace_files = {}
-    for root, _, files in os.walk(os.path.join(data_directory, investment_platform.name)):
-        for investment_snapshot in files:
-            match = investment_platform.filename_regexp.search(investment_snapshot)
-            if match:
-                report_date = datetime.date.fromisoformat(f"{match.group('year')}-{match.group('month')}-{match.group('day')}")
-                marketplace_files[report_date] = os.path.join(root, investment_snapshot)
-    return marketplace_files
-
-
-def collect_investment_data(data_directory, investment_platforms):
-    data_files = {}
-    for investment_platform in investment_platforms:
-        data_files[investment_platform] = read_marketplace_files(data_directory, investment_platform)
-    return data_files
 
 
 def aggregate_investment_data(investment_files):
