+++ test/test_p2p.py
@@ -1,9 +1,5 @@
 import datetime
-from unittest.mock import patch
-from marketplace import mintos
-from marketplace import iuvo
-from p2p import get_latest_report_date, filter_investment_files_by_newest_date, \
-    read_marketplace_files
+from p2p import get_latest_report_date, filter_investment_files_by_newest_date
 
 
 def test_get_latest_report_date():
@@ -16,42 +12,6 @@ def test_get_latest_report_date():
         datetime.date.fromisoformat("2021-01-01") : "path6.xls",
     }
     assert datetime.date.fromisoformat("2021-12-01") == get_latest_report_date(input_data)
-
-
-@patch('p2p.os.walk')
-def test_read_marketplace_files_iuvo(os_walk):
-    os_walk.return_value = [
-        ('/iuvo', ('',), (
-            'MyInvestments-20201130.xlsx',
-            'MyInvestments-20201231.xlsx',
-            '20201130-current-investments.xlsx',
-            '20210630-current-investments.xlsx',
-            'OtherFile.xlsx',
-        ))
-    ]
-    output_data = {
-        datetime.date.fromisoformat("2020-11-30"): "/iuvo/MyInvestments-20201130.xlsx",
-        datetime.date.fromisoformat("2020-12-31"): "/iuvo/MyInvestments-20201231.xlsx",
-    }
-    assert output_data == read_marketplace_files("dummy/folder", iuvo.META_DATA)
-
-
-@patch('p2p.os.walk')
-def test_read_marketplace_files_mintos(os_walk):
-    os_walk.return_value = [
-        ('/mintos', ('',), (
-            'MyInvestments-20201130.xlsx',
-            'MyInvestments-20201231.xlsx',
-            '20201130-current-investments.xlsx',
-            '20210630-current-investments.xlsx',
-            'OtherFile.xlsx',
-        ))
-    ]
-    output_data = {
-        datetime.date.fromisoformat("2020-11-30"): "/mintos/20201130-current-investments.xlsx",
-        datetime.date.fromisoformat("2021-06-30"): "/mintos/20210630-current-investments.xlsx",
-    }
-    assert output_data == read_marketplace_files("dummy/folder", mintos.META_DATA)
 
 
 def test_filter_investment_files_by_newest_date():
